main/requirements.txt
discord.py==0.16.12
aiohttp
google-api-python-client
google-auth
google-auth-httplib2
httplib2
oauth2client